Move scene from ITMTrackingState -> ITMRenTracker.
ITMTrackingState isn't templated, so we can't store a scene pointer in it
without hard-coding the template parameters. This prevents external code
from using custom voxel types. By moving the scene pointer into
ITMRenTracker, which is both templated and the only place in which the
pointer is used, we solve the problem.