Better handling of block allocation failures
If the allocation fails due to not enough blocks available mark the entry as not visible (undoing what buildHashAllocAndVisibleTypePP did)
This prevents crashes later due to the out of bounds access caused by the presence of invalid indices in the visible entries list. 
The invalid indices are inserted during the reduction in the buildVisibleList function.