Added requiresPointCloudRendering() method to the tracker interface.
The prepare method of the tracking controller now does not perform the rendering if there is no need for a raycasted point cloud. The track method in the controller is also updated accordingly.
Used when starting a reconstruction with a ground truth trajectory: before this change the first pose was always the identity, now the pose is the one stored in the GT file.